Refactor layout
