+++ src/layouts/default-layout/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./default-layout"
